Rename phone-masks.ts to form-input-masks.ts and update imports, add tax id mask
