+++ packages/webcomponents/src/components/business-form/business-owners/business-owners.tsx
@@ -1,6 +1,6 @@
 import { Component, Host, h, Prop, State } from '@stencil/core';
 import { FormController } from '../../form/form';
-import { PHONE_MASKS } from '../../../utils/phone-masks';
+import { PHONE_MASKS } from '../../../utils/form-input-masks';
 
 class BusinessOwner {
   name: string = '';
