+++ packages/webcomponents/src/utils/form-input-masks.ts
@@ -1,3 +1,5 @@
 export const PHONE_MASKS = {
   US: '(000) 000-0000',
 };
+
+export const TAX_ID_MASKS = { US: '00-000000000' };
